pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ragplus"
version = "0.2.0"
description = "A lightweight toolkit for building simple and fast RAG (Retrieval-Augmented Generation) pipelines."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "muthuvel vishwa", email = "muthuvelvishwa2002@gmail.com" }
]
keywords = ["rag", "retrieval", "embeddings", "vector search", "nlp", "llm"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.22.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "pdfplumber>=0.9.0",
    "python-docx>=0.8.11",
    "rank-bm25>=0.2.2",
    "nltk>=3.8",
    "chardet>=5.0.0"
]


[project.urls]
Homepage = "https://github.com/vishwa7890/ragplus"
Repository = "https://github.com/vishwa7890/ragplus"

[project.scripts]
ragplus-demo = "ragplus.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ragplus*"]
exclude = ["ragplus_data*", "tests*", "examples*"]
